move tx summary computation to separate transaction/summary crate this simplifies building for hardware wallets as we no longer need to include transaction/{core,extra}
