+++ core/Cargo.toml
@@ -26,7 +26,7 @@ slip10_ed25519 = { version = "0.1", optional = true }
 tiny-bip39 = { version = "1.0", optional = true }
 zeroize = { version = "1.5", default-features = false }
 
-mc-core-types = { path = "./types" }
+mc-core-types = { path = "./types", default-features = false }
 mc-crypto-hashes = { path = "../crypto/hashes", default-features = false }
 mc-crypto-keys = { path = "../crypto/keys", default-features = false }
