+++ crypto/ring-signature/Cargo.toml
@@ -32,7 +32,7 @@ mc-crypto-hashes = { path = "../../crypto/hashes", default-features = false }
 mc-crypto-keys = { path = "../../crypto/keys", default-features = false }
 mc-util-from-random = { path = "../../util/from-random" }
 mc-util-repr-bytes = { path = "../../util/repr-bytes", default-features = false }
-mc-util-serial = { path = "../../util/serial", optional = true }
+mc-util-serial = { path = "../../util/serial", optional = true, default_features = false }
 
 # Enable all default features not known to break code coverage builds
 proptest = { version = "1.1", default-features = false, features = ["default-code-coverage"], optional = true }
