+++ transaction/extra/src/lib.rs
@@ -16,7 +16,6 @@ mod memo;
 mod signed_contingent_input;
 mod tx_out_confirmation_number;
 mod tx_out_gift_code;
-mod tx_summary_unblinding;
 mod unsigned_tx;
 
 pub use memo::{
@@ -32,10 +31,6 @@ pub use signed_contingent_input::{
 };
 pub use tx_out_confirmation_number::TxOutConfirmationNumber;
 pub use tx_out_gift_code::TxOutGiftCode;
-pub use tx_summary_unblinding::{
-    verify_tx_summary, TransactionEntity, TxOutSummaryUnblindingData, TxSummaryStreamingVerifier,
-    TxSummaryUnblindingData, TxSummaryUnblindingReport,
-};
 pub use unsigned_tx::UnsignedTx;
 
 // Re-export this to help the exported macros work
