--- transaction/extra/src/tx_summary_unblinding/mod.rs
@@ -1,11 +1,0 @@
-// Copyright (c) 2018-2022 The MobileCoin Foundation
-
-mod data;
-mod error;
-mod report;
-mod verifier;
-
-pub use data::{TxOutSummaryUnblindingData, TxSummaryUnblindingData};
-pub use error::Error;
-pub use report::{TransactionEntity, TxSummaryUnblindingReport};
-pub use verifier::{verify_tx_summary, TxSummaryStreamingVerifier};
