+++ transaction/signer/Cargo.toml
@@ -33,6 +33,7 @@ mc-crypto-ring-signature = { path = "../../crypto/ring-signature" }
 mc-crypto-ring-signature-signer = { path = "../../crypto/ring-signature/signer" }
 mc-transaction-core = { path = "../../transaction/core" }
 mc-transaction-extra = { path = "../../transaction/extra" }
+mc-transaction-summary = { path = "../../transaction/summary" }
 mc-util-repr-bytes = { path = "../../util/repr-bytes", default-features = false }
 mc-util-serial = { path = "../../util/serial", default-features = false }
