+++ util/zip-exact/Cargo.toml
@@ -8,5 +8,8 @@ license = "Apache-2.0"
 readme = "README.md"
 rust-version = { workspace = true }
 
+[features]
+default = ["serde"]
+
 [dependencies]
-serde = { version = "1.0", default-features = false }
+serde = { version = "1.0", default-features = false, optional = true }
